refactor(strkern): one header — the sibling scan folds back in and the escapers get the block loop
The owner's rule (2026-09-10) is that ALL SIMD string kernels live in src/infra/strkern.h. Lane M
could not follow it: `findByteset`'s tail was the oracle, so the escapers could not call the one
header without paying 5x, and M shipped `src/infra/strkern_find.h` instead — recording in that
file's own closing note that "if strkern.h grows a shipped tail beside its oracle, this file
collapses into a call to it and the SIMD path comes along for free". The previous commit grew the
tail. This commit collapses the file.

  * `appendCleanRun` moves into strkern.h as section 4, unchanged in body except for the scan it
    calls, with M's attributions and reasoning kept verbatim.
  * `findBytesetRun` does not move: with the tail fixed it WAS `findByteset_scalar`, character for
    character. One fewer name for one function.
  * src/infra/strkern_find.h deleted; src/serialize.h and src/infra/jsonesc.h include strkern.h.
  * `rg -n 'arm_neon.h|immintrin.h' src/` now lists strkern.h plus exactly the four pre-existing
    infra headers (dynamic_map.hpp, radixSort.h, sparseCsr.h, fixedStr.h). No sixth home.

THE SCAN UNDER appendCleanRun IS NOW `findByteset` — block loop and all — because it measured
better, on the inputs ripwire actually escapes rather than on a synthetic band. A scratch
RIPWIRE_ESCAPE_TRACE build (reverted, never committed) captured every escapeXml argument of a warm
`--top-k=100000` map; the traces were replayed end to end through escapeXml, best of 9, four
independent process runs, box load 24-33. Milliseconds for the whole trace:

  corpus    calls    median len   per-byte switch   + scalar scan   + findByteset
  ripwire   44,670       11        1.60-2.00         0.81-1.01       0.71-0.96   -15%
  go       340,814        8        6.85-7.62         4.57-4.85       4.74-5.08    +1%
  django   148,681       17        8.00-9.12         3.80-4.30       2.97-3.11   -23%

Two win, one ties inside the noise band, none loses — so the block loop ships and no caller has to
choose a scan. go is the tie because its median escaped string is 8 bytes and 82% of its calls never
reach a 16-byte block; that is the same fact that made the old tail so expensive.

WHAT IS NOT CLAIMED. escapeXml is ~1% of a warm map on these corpora (0.82-1.18% by `sample`), so a
15-23% cut in it is ~0.2% of the verb. The interleaved whole-verb A/B was run anyway (12 arms a side
per corpus, user+sys) and resolves nothing: ripwire tree 0.110 / 0.110 s median, django 0.540 /
0.530, go 0.940 / 0.940. The `sample` arm M's report prescribes was also run and is REPORTED AS
INCONCLUSIVE rather than dressed up: at this box load the escaper draws 27-58 samples out of
3.3k-7.1k busy per arm, which cannot resolve a 20% change in a 1% site. The trace replay is the
instrument; the sampler is not, here.

Byte-identical, 21/21: 3 corpora (frozen git-archive of this tree, go, django) x 7 surfaces
(--top-k=100000 XML and JSON, --for XML and JSON, --grep, --pack-task, --lint) — stdout, stderr and
exit code all cmp-equal against the pre-change binary. Determinism (two runs cmp-equal) and
`xmllint --noout` clean. emitescapecheck ALL PASS (222,682 inputs, MUT arm sees 168,423 disagree);
strkerncheck PASS (14 arms NEON, 14 arms AVX2 under Rosetta 2, mutation reds 9).
